Add third parameter to button template tag
Many apps depend on a `STATIC_URL` template variable to prefix the
locations of images to avoid repetition or enable something like a
CDN to be used. There needed to be a way to follow the practice
with socialregistration.

This commit introduces a second optional parameter to the `button`
base template tag. It expects an unquoted context variable and the
contents of that variable will be prepended to the first parameter.